+++ backend/apps/accounts/forms/newsletter.py
@@ -2,70 +2,9 @@ from crispy_forms.bootstrap import PrependedText
 from crispy_forms.helper import FormHelper
 from crispy_forms.layout import Column, HTML, Layout, Row, Submit
 from django import forms
-from django.contrib.auth.forms import AuthenticationForm, UserCreationForm, UsernameField
-from django.db import transaction
 from django.utils.translation import gettext_lazy as _
 
-from .modelss import Newsletter, User
-
-
-class HospitalSignUpForm(UserCreationForm):
-    # add more query fields
-
-    class Meta(UserCreationForm.Meta):
-        model = User
-        fields = []  # ['email']
-
-    @transaction.atomic
-    def save(self):
-        user = super().save(commit=False)
-        user.is_hospital = True
-        user.save()
-        return user
-
-
-class StudentEmailForm(forms.ModelForm):
-    class Meta(UserCreationForm.Meta):
-        model = User
-        fields = ["email"]
-
-    @transaction.atomic
-    def save(self):
-        user = super().save(commit=False)
-        user.username = user.email
-        user.save()
-        return user
-
-
-class HospitalEmailForm(forms.ModelForm):
-    class Meta(UserCreationForm.Meta):
-        model = User
-        fields = ["email"]
-
-    @transaction.atomic
-    def save(self):
-        user = super().save(commit=False)
-        user.username = user.email
-        user.save()
-        return user
-
-
-class StudentSignUpForm(UserCreationForm):
-    # add more query fields
-
-    class Meta(UserCreationForm.Meta):
-        model = User
-
-    @transaction.atomic
-    def save(self):
-        user = super().save(commit=False)
-        user.is_student = True
-        user.save()
-        return user
-
-
-class CustomAuthenticationForm(AuthenticationForm):
-    username = UsernameField(label=_("E-Mail"), widget=forms.TextInput(attrs={"autofocus": True}))
+from apps.accounts.modelss import Newsletter
 
 
 class BaseNewsletterForm(forms.ModelForm):
@@ -129,7 +68,3 @@ class NewsletterViewForm(BaseNewsletterForm):
             self.fields[f].disabled = True
             self.fields[f].required = False
         self.helper.form_tag = False
-
-
-class TestMailForm(forms.Form):
-    email = forms.EmailField()
