+++ backend/apps/accounts/forms/test_mail.py
@@ -1,0 +1,5 @@
+from django import forms
+
+
+class TestMailForm(forms.Form):
+    email = forms.EmailField()
